requirements.txt
Flask>=2.3.3
Pillow>=9.5,<12.0
stegano>=0.11.5
Werkzeug>=2.3.7
